bug al anular servicio si no tiene pago resuelto

+++ lib/payments_and_services/invoices/anularService.ts
@@ -25,44 +25,45 @@ export async function anularService(service_id: string) {
 		throw new Error('Error al anular el servicio' + paymentError.message);
 	}
 
-	const { data: dataPrepayment } = await supabase
-		.from('prepayments')
-		.select('amount')
-		.eq('pay_id', paymentData?.[0].id);
+	if (!(paymentData.length === 0)) {
+		const { data: dataPrepayment } = await supabase
+			.from('prepayments')
+			.select('amount')
+			.eq('pay_id', paymentData?.[0].id);
 
-	let amount = 0;
+		let amount = 0;
 
-	if (!(dataPrepayment?.length === 0)) {
-		amount = dataPrepayment?.[0].amount;
+		if (!(dataPrepayment?.length === 0)) {
+			amount = dataPrepayment?.[0].amount;
+		}
+
+		let devolution = data[0].deuda + (data[0].monto + amount);
+		await supabase.from('prepayments').delete().eq('pay_id', paymentData?.[0].id);
+
+		if (devolution > 0) {
+			data[0].payment_id.forEach(async (payment_id: string) => {
+				const { data } = await supabase
+					.from('payments')
+					.select('monto_ref, client_id')
+					.eq('id', payment_id);
+
+				if (devolution > data?.[0].monto_ref) {
+					devolution -= data?.[0].monto_ref;
+					await supabase.from('prepayments').insert({
+						amount: data?.[0].monto_ref,
+						client_id: data?.[0].client_id,
+						pay_id: payment_id,
+					});
+				} else if (devolution > 0) {
+					await supabase.from('prepayments').insert({
+						amount: devolution,
+						client_id: data?.[0].client_id,
+						pay_id: payment_id,
+					});
+				}
+			});
+		}
 	}
-
-	let devolution = data[0].deuda + (data[0].monto + amount);
-	await supabase.from('prepayments').delete().eq('pay_id', paymentData?.[0].id);
-
-	if (devolution > 0) {
-		data[0].payment_id.forEach(async (payment_id: string) => {
-			const { data } = await supabase
-				.from('payments')
-				.select('monto_ref, client_id')
-				.eq('id', payment_id);
-
-			if (devolution > data?.[0].monto_ref) {
-				devolution -= data?.[0].monto_ref;
-				await supabase.from('prepayments').insert({
-					amount: data?.[0].monto_ref,
-					client_id: data?.[0].client_id,
-					pay_id: payment_id,
-				});
-			} else if (devolution > 0) {
-				await supabase.from('prepayments').insert({
-					amount: devolution,
-					client_id: data?.[0].client_id,
-					pay_id: payment_id,
-				});
-			}
-		});
-	}
-
 	await supabase.from('service_receivable').update({ estado: false }).eq('id', service_id);
 
 	return 'Servicio anulado exitosamente';
